read_han/pyproject.toml
[build-system]
requires = [ "setuptools" ]
build-backend = "setuptools.build_meta"

[project]
name = "read_han"
version = "0.1"
dependencies = [
    "pyserial"
]

[project.scripts]
read_han = "app:main"
